Demonstrate two approaches with Redux Persist
Before the example loaded the page immediately, when it could not
possibly have access to persisted state. This leads to a fairly abrupt
change when the client takes over, which is not desirable for an actual
app, but does help demonstrate what is actually going on.

Now the second page uses a different approach, choosing to display a
loading message where any the page could change after rehydration.